refactor: move Magic to format
