+++ src/format/mod.rs
@@ -2,3 +2,4 @@
 
 pub(crate) mod aes;
 pub(crate) mod flags;
+pub(crate) mod magic;
